requirements.txt
Flask>=2.3.0
Flask-Cors>=3.0.0

langchain-core>=1.0.0,<2.0.0

langchain>=0.2.0,<1.0.0

langchain-community>=0.2.0,<1.0.0

langchain-google-genai>=3.0.0,<4.0.0

langchain-huggingface>=0.0.3,<0.1.0

google-generativeai>=0.7.0,<1.0.0

Vectorstore and embeddings

faiss-cpu>=1.7.0
sentence-transformers>=2.6.0